[refactor] ModelRetrieveService->OfferRetrieveService
