--- src/main/java/com/moddy/server/controller/model/dto/HairServiceOfferIdDTO.java
@@ -1,7 +1,0 @@
-package com.moddy.server.controller.model.dto;
-
-public record HairServiceOfferIdDTO(
-        Long designerId,
-        Long applicationId
-) {
-}
